Improved content.js
Made content.js much more efficient.

Blocks were previously displaying with HTML as plaintext in popup because each segment of the block was being saved as a string and += to the innerHTML of the div, but they are now consolidated into one single string that is set to the innerHTML of the div.